Upgrade proto to add error code field of Status
